refactor💥: models.utils > model.misc
Naming it "utils" messed with "client.utils" for some reason.
