+++ dis_snek/models/__init__.py
@@ -1,3 +1,3 @@
 from .discord import *
 from .snek import *
-from .utils import *
+from .misc import *
