+++ dis_snek/models/misc/iterator.py
