+++ docs/src/API Reference/models/.pages
@@ -1,0 +1,4 @@
+nav:
+    - Discord
+    - Snek
+    - Misc
