+++ docs/src/API Reference/models/Misc/iterator.md
@@ -1,0 +1,1 @@
+::: dis_snek.models.misc.iterator
