--- docs/src/API Reference/models/Utils/iterator.md
@@ -1,1 +1,0 @@
-::: dis_snek.models.utils.iterator
