+++ docs/src/API Reference/models/index.md
@@ -8,5 +8,5 @@ Within these pages, you will find a list of all available models within Snek.
 - [Snek](Snek)
     - Models for library constructs
 
-- [Utils](Utils)
+- [Misc](Misc)
     - Models for constructs that fit neither category
